update_DNS_records/requirements.txt
cloudflare==3.1.0
requests==2.32.3
argparse==1.4.0
pydantic==2.9.2
httpx==0.27.2
pyyaml==6.0.2